VistaAspects/VistaConversion:   ADD: specialization for VistaColor
git-svn-id: https://svn.rwth-aachen.de/repos/vrgroup-svn/projects/Vista/trunk@29111 33c38b33-09d2-4561-8d4d-1bfa2f58f885

+++ VistaCoreLibs/Tests/VistaTestingUtils.h
@@ -58,7 +58,7 @@ namespace VistaTestingRandom
 		dRand *= S_nRandRange;
 		return dRand;
 	}
-	inline double RandomFloat()
+	inline float RandomFloat()
 	{
 		return (float)RandomDouble();
 	}
